requirements.txt
fastapi
uvicorn
tensorflow
numpy
pydantic
Pillow